+++ CHANGELOG.rst
@@ -4,7 +4,8 @@ Changelog
 2020-11-16 (version 1.15):
     * quick labels: CTRL + right mouse button opens map menu even when the feature modify map tool is activates
     * source visibility update can be run on entire time series or (new and faster) for the next time steps only
-    * time series table can keep focus on the date range that is visualized in the map windows
+    * added "follow current date" option to time series table to keep focus on the map window date range
+    * added wildcard +  regular expression filter to time series table
     * smaller bug fixes and improvements of workflow
 
 2020-11-06 (version 1.14):